requirements.txt
mysql-connector-python==8.0.21
onetimepad==1.4
playsound==1.2.2
pyaudio==0.2.11
pyttsx3==2.90
SpeechRecognition==3.8.1
sqlparse==0.3.1
tabulate==0.8.7